requirements.txt
uvicorn

pyannote.audio
openai-whisper
torchaudio
torch
python-Levenshtein
pydu
speechbrain
silero-vad
seaborn

psycopg2-binary
spacy
transformers
statsmodels
ipython
datasets
python-dotenv

boto3